requirements.txt
attrs==17.2.0
Automat==0.6.0
backports-abc==0.5
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
constantly==15.1.0
coreapi==2.3.1
coreapi-cli==1.0.6
coreschema==0.0.4
Django==1.11.5
django-filter==1.0.4
djangorestframework==3.6.4
gevent==1.2.2
greenlet==0.4.12
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
MySQL-python==1.2.5
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.11.0
tornado==4.5.2
Twisted==17.9.0
uritemplate==3.0.0
urllib3==1.22
zope.interface==4.4.3
